# Optimize LEB128 data reading